Projectgen

Former-commit-id: 20eae4dfb42fa8e0a6d08dfa5ba7fbf5e5cc3106
Former-commit-id: a885b5c82f99a279fd32c04c9c476b467dd4df60 [formerly b5d9e4652f5e800ec2fe37ad0182f74ce5cf7fd4]
Former-commit-id: c9fc7ad8ced48bd06e88588d947cdcbc1b4b8c9a

+++ modules/util/source/hip/util/system.d
@@ -21,7 +21,7 @@ pure nothrow string sanitizePath(string path)
 {
     string ret = new string(path.length);
 
-    for(uint i = 0; i < path.length; i++)
+    for(size_t i = 0; i < path.length; i++)
     {
         version(Windows)
         {
@@ -42,7 +42,7 @@ pure nothrow string sanitizePath(string path)
 }
 pure nothrow bool isPathUnixStyle(string path)
 {
-    for(int i = 0; i < path.length; i++)
+    for(size_t i = 0; i < path.length; i++)
         if(path[i] == '/')
             return true;
     return false;
